refactor: rename files to fit conventions
