+++ packages/article/src/article-lead-asset/article-lead-asset.js
